Move CompileState
