+++ src/main/java/io/github/ngspace/hudder/compilers/UnaccessableTestCompiler.java
@@ -11,6 +11,7 @@ public class UnaccessableTestCompiler extends ATextCompiler {
 	}
 
 	@Override public Object getVariable(String key) throws CompileException {
+		//I made a typo... I am keeping this here...
 		return "Advanchment unlocked: How did we get here?";
 	}
