+++ src/main/java/io/github/ngspace/hudder/compilers/utils/CompileState.java
@@ -1,11 +1,9 @@
-package io.github.ngspace.hudder.methods;
+package io.github.ngspace.hudder.compilers.utils;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import io.github.ngspace.hudder.compilers.utils.CompileException;
-import io.github.ngspace.hudder.compilers.utils.CompileResult;
 import io.github.ngspace.hudder.config.ConfigManager;
 import io.github.ngspace.hudder.methods.elements.AUIElement;
