+++ src/main/java/io/github/ngspace/hudder/methods/methods/ItemStackMethods.java
@@ -4,8 +4,8 @@ import static io.github.ngspace.hudder.Hudder.ins;
 
 import io.github.ngspace.hudder.compilers.ATextCompiler;
 import io.github.ngspace.hudder.compilers.utils.CompileException;
+import io.github.ngspace.hudder.compilers.utils.CompileState;
 import io.github.ngspace.hudder.config.ConfigInfo;
-import io.github.ngspace.hudder.methods.CompileState;
 import io.github.ngspace.hudder.methods.MethodValue;
 import io.github.ngspace.hudder.methods.elements.ItemElement;
 import net.minecraft.entity.player.PlayerInventory;
