+++ src/main/java/io/github/ngspace/hudder/v2runtime/V2Runtime.java
@@ -3,7 +3,7 @@ package io.github.ngspace.hudder.v2runtime;
 import java.util.Arrays;
 
 import io.github.ngspace.hudder.compilers.utils.CompileException;
-import io.github.ngspace.hudder.methods.CompileState;
+import io.github.ngspace.hudder.compilers.utils.CompileState;
 import io.github.ngspace.hudder.methods.MethodHandler;
 import io.github.ngspace.hudder.v2runtime.runtime_elements.AV2RuntimeElement;
