+++ src/main/java/io/github/ngspace/hudder/v2runtime/runtime_elements/VariableV2RuntimeElement.java
@@ -1,7 +1,7 @@
 package io.github.ngspace.hudder.v2runtime.runtime_elements;
 
 import io.github.ngspace.hudder.compilers.utils.CompileException;
-import io.github.ngspace.hudder.methods.CompileState;
+import io.github.ngspace.hudder.compilers.utils.CompileState;
 import io.github.ngspace.hudder.v2runtime.AV2Compiler;
 import io.github.ngspace.hudder.v2runtime.values.V2Value;
