+++ src/main/java/io/github/ngspace/hudder/v2runtime/runtime_elements/WhileV2RuntimeElement.java
@@ -1,8 +1,8 @@
 package io.github.ngspace.hudder.v2runtime.runtime_elements;
 
 import io.github.ngspace.hudder.compilers.utils.CompileException;
+import io.github.ngspace.hudder.compilers.utils.CompileState;
 import io.github.ngspace.hudder.config.ConfigInfo;
-import io.github.ngspace.hudder.methods.CompileState;
 import io.github.ngspace.hudder.v2runtime.AV2Compiler;
 import io.github.ngspace.hudder.v2runtime.values.V2Value;
